arm: dts: Set initial gpu powerlevel to 4
Powerlevel 4 equals to 320MHz. The same idle value is used in the stock kernel.

Raising it minimizes lag at the beginning of animations as the gpu is not at its lowest
frequency when idle. Tradeoff: A little battery life.

The value of 6 (130MHz) worked well on LOS 14 but causes some UI lag on 15.

Signed-off-by: Planet-X <planetx.dev@gmail.com>